move to appsembler/usage with custom AppConfig
