+++ openedx/core/djangoapps/appsembler/usage/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'openedx.core.djangoapps.appsembler.usage.apps.AppsemblerUsageAppConfig'
