+++ openedx/core/djangoapps/appsembler/usage/apps.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import, unicode_literals
+
+from django.apps import AppConfig
+
+
+class AppsemblerUsageAppConfig(AppConfig):
+    name = 'openedx.core.djangoapps.appsembler.usage'
+    label = 'appsembler_usage'
+
